Updated to EXILED 7.2
Note that this is an old version, and likely won't work now.
This has been sitting on my computer for a while and I want to publish it.
I won't be updating it to the latest version unless I feel like it.
Also, intercom stuff has been removed, but could be added back.